Initial Commit (from github Desktop)
I am following this advice on youtube: https://www.youtube.com/watch?v=qBcsczLJkUY to submit an existing project to github.  I'd previously developed this (unfinished) project in another location, calling it "CheckoutBelt".  It simply uses a cube to act as a checkout counter.  I am pushing prefabs along the belt using RigidBody.velocity, and I'm using TextmeshPro to hang a "PriceTag" above the prefab.  It's coming along nicely, but there is a lot of work to be done before it's finished!